refactor: move getSizeOfValue from src/btree/ to src/json.ts
